refactor(activerecord): extract Validations surface from Base to validations.ts
Moves the ActiveRecord::Validations methods out of base.ts into
validations.ts where Rails keeps them. Wires back onto Base via the
existing extend/include plumbing.

Class methods (extend via Validations.ClassMethods):
- validates — the AR override that routes presence/absence/length/
  numericality to AR-specific validator classes, then delegates
  remaining rules (inclusion/exclusion/format/...) through to
  ActiveModel's validates via a late-bound super (_setSuperValidates).
- validatesAssociated
- validatesUniqueness

Instance methods (wired via include):
- readAttributeForValidation
- validate (already existed as module function; now also on Base)
- customValidationContext (already existed; now also on Base)

extractShared helper moved into validations.ts alongside validates.

Matches Rails' layout: ActiveRecord::Validations lives in
validations.rb with a ClassMethods mixin. Our validations.ts now
holds the same split.

Pure delegation — no behavior change. 17800 tests pass, DX types
pass, guides typecheck pass. api:compare steady.
