+++ packages/activerecord/src/validations.ts
@@ -6,13 +6,20 @@
  * and overrides save/valid? to run validations with context awareness.
  */
 import { ActiveRecordError } from "./errors.js";
+import { AbsenceValidator } from "./validations/absence.js";
+import { AssociatedValidator } from "./validations/associated.js";
+import { LengthValidator } from "./validations/length.js";
+import { NumericalityValidator } from "./validations/numericality.js";
+import { PresenceValidator } from "./validations/presence.js";
 
 // Re-export all validators matching Rails' require at bottom of validations.rb
-export { AbsenceValidator } from "./validations/absence.js";
-export { AssociatedValidator } from "./validations/associated.js";
-export { LengthValidator } from "./validations/length.js";
-export { NumericalityValidator } from "./validations/numericality.js";
-export { PresenceValidator } from "./validations/presence.js";
+export {
+  AbsenceValidator,
+  AssociatedValidator,
+  LengthValidator,
+  NumericalityValidator,
+  PresenceValidator,
+};
 export { UniquenessValidator } from "./validations/uniqueness.js";
 
 /**
@@ -124,3 +131,173 @@ export function performValidations(
   if (options?.validate === false) return true;
   return this.isValid(options?.context);
 }
+
+/**
+ * Mirrors: ActiveModel::Validations#read_attribute_for_validation
+ *
+ * Rails aliases this to `send`, so calling it with an association name
+ * returns the association target (loaded records). We resolve from
+ * association caches first, falling back to readAttribute for regular
+ * columns.
+ */
+export function readAttributeForValidation(this: any, attribute: string): unknown {
+  const cached = this._cachedAssociations?.get?.(attribute);
+  if (cached !== undefined) return cached;
+  const preloaded = this._preloadedAssociations?.get?.(attribute);
+  if (preloaded !== undefined) return preloaded;
+  const proxy = this._collectionProxies?.get?.(attribute);
+  if (
+    proxy &&
+    (proxy.loaded === true || (Array.isArray(proxy.target) && proxy.target.length > 0))
+  ) {
+    return proxy.target;
+  }
+  if (typeof this.association === "function") {
+    try {
+      const assoc = this.association(attribute);
+      if (assoc?.loaded === true && assoc.target !== undefined) return assoc.target;
+    } catch {
+      // Not an association — fall through
+    }
+  }
+  return this.readAttribute(attribute);
+}
+
+// ---------------------------------------------------------------------------
+// Class methods — Mirrors ActiveRecord::Validations::ClassMethods.
+// Wired onto Base via extend(Base, Validations.ClassMethods) in base.ts.
+// ---------------------------------------------------------------------------
+
+// Options passed alongside any validator: on/if/unless/strict plus
+// allowNil/allowBlank that are shared across all validators.
+function extractShared(rules: Record<string, unknown>): Record<string, unknown> {
+  const shared: Record<string, unknown> = {};
+  if (rules.on !== undefined) shared.on = rules.on;
+  if (rules.if !== undefined) shared.if = rules.if;
+  if (rules.unless !== undefined) shared.unless = rules.unless;
+  if (rules.strict) shared.strict = rules.strict;
+  if (rules.allowNil !== undefined) shared.allowNil = rules.allowNil;
+  if (rules.allowBlank !== undefined) shared.allowBlank = rules.allowBlank;
+  return shared;
+}
+
+/**
+ * Route AR-specific rules (presence/absence/length/numericality) through AR
+ * validator classes that add association/column awareness, and delegate the
+ * rest (inclusion/exclusion/format/...) to ActiveModel's `validates`.
+ *
+ * Mirrors: ActiveRecord::Validations::ClassMethods#validates (the override
+ * over ActiveModel::Validations::ClassMethods#validates).
+ */
+export function validates(
+  this: {
+    validatesWith(validatorClass: unknown, opts: Record<string, unknown>): void;
+    // Model.prototype.validates reached via the parent prototype chain.
+  },
+  attribute: string,
+  rules: Record<string, unknown>,
+): void {
+  const arRules = { ...rules };
+  const shared = extractShared(arRules);
+  const { allowNil: sharedAllowNil, allowBlank: sharedAllowBlank, ...sharedRest } = shared;
+
+  const buildOpts = (opts: Record<string, unknown>) => ({
+    ...opts,
+    attributes: [attribute],
+    ...sharedRest,
+    ...(opts.allowNil === undefined && sharedAllowNil !== undefined
+      ? { allowNil: sharedAllowNil }
+      : {}),
+    ...(opts.allowBlank === undefined && sharedAllowBlank !== undefined
+      ? { allowBlank: sharedAllowBlank }
+      : {}),
+  });
+
+  if (arRules.presence) {
+    const opts = arRules.presence === true ? {} : (arRules.presence as Record<string, unknown>);
+    delete arRules.presence;
+    this.validatesWith(PresenceValidator, buildOpts(opts));
+  }
+  if (arRules.absence) {
+    const opts = arRules.absence === true ? {} : (arRules.absence as Record<string, unknown>);
+    delete arRules.absence;
+    this.validatesWith(AbsenceValidator, buildOpts(opts));
+  }
+  if (arRules.length) {
+    const opts = arRules.length as Record<string, unknown>;
+    delete arRules.length;
+    this.validatesWith(LengthValidator, buildOpts(opts));
+  }
+  if (arRules.numericality) {
+    const opts =
+      arRules.numericality === true ? {} : (arRules.numericality as Record<string, unknown>);
+    delete arRules.numericality;
+    this.validatesWith(NumericalityValidator, buildOpts(opts));
+  }
+  // Delegate remaining rules (inclusion/exclusion/format/...) to ActiveModel's validates.
+  const hasRemaining = Object.keys(arRules).some(
+    (k) => !["on", "if", "unless", "strict", "allowNil", "allowBlank"].includes(k),
+  );
+  if (hasRemaining) {
+    // `super.validates` — reach the parent (Model) prototype's version.
+    _parentValidates?.call(this, attribute, arRules);
+  }
+}
+
+// Late-bound reference to Model.prototype.validates — registered by
+// Base to break the circular-import chain.
+let _parentValidates:
+  | ((this: unknown, attribute: string, rules: Record<string, unknown>) => void)
+  | null = null;
+
+/** @internal Called by Base to register Model's validates as the super. */
+export function _setSuperValidates(
+  fn: (this: unknown, attribute: string, rules: Record<string, unknown>) => void,
+): void {
+  _parentValidates = fn;
+}
+
+/**
+ * Validates that all named associations are themselves valid.
+ *
+ * Mirrors: ActiveRecord::Validations::ClassMethods#validates_associated
+ */
+export function validatesAssociated(
+  this: { validatesWith(vc: unknown, opts: Record<string, unknown>): void },
+  ...args: (string | Record<string, unknown>)[]
+): void {
+  const last = args[args.length - 1];
+  const opts =
+    typeof last === "object" && last !== null ? (args.pop() as Record<string, unknown>) : {};
+  for (const name of args as string[]) {
+    this.validatesWith(AssociatedValidator, { ...opts, attributes: [name] });
+  }
+}
+
+/**
+ * Register a deferred uniqueness validation to run on save (since uniqueness
+ * requires a DB round-trip, it's kept off the synchronous validator chain).
+ *
+ * Mirrors: validates uniqueness: true / ActiveRecord::Validations::ClassMethods#validates_uniqueness_of
+ */
+export function validatesUniqueness(
+  this: unknown,
+  attribute: string,
+  options: { scope?: string | string[]; message?: string; conditions?: (this: any) => any } = {},
+): void {
+  const klass = this as { _asyncValidations?: Array<unknown> };
+  if (!Object.prototype.hasOwnProperty.call(klass, "_asyncValidations")) {
+    klass._asyncValidations = [...(klass._asyncValidations ?? [])];
+  }
+  (klass._asyncValidations as Array<unknown>).push({ attribute, options });
+}
+
+/**
+ * Module methods wired onto Base as static methods via `extend()` in base.ts.
+ * Mirrors Rails' `ActiveRecord::Validations::ClassMethods` / `ActiveSupport::Concern#ClassMethods`.
+ */
+export const ClassMethods = {
+  validates,
+  validatesAssociated,
+  validatesUniqueness,
+};
